do not reuse forks in surefire to reset peersim.Comfiguration
